+++ packages/react-native-gesture-handler/src/__tests__/api_v3.test.tsx
@@ -256,14 +256,18 @@ describe('[API v3] Components', () => {
 
       await act(flushImmediate);
 
-      const nativeDetector = getNativeDetector(UNSAFE_getAllByType);
       const scrollViewResponder = getScrollViewResponder(UNSAFE_getAllByType);
+      const button = screen.getByTestId('pressable');
 
+      // Default Pressable delegates to the native-button Touchable, which marks
+      // the tap as RNGH-handled via the button's capture handler.
       expect(scrollViewResponder).toBeDefined();
       expect(
         scrollViewResponder?.props.onStartShouldSetResponderCapture()
       ).toBe(false);
-      expect(nativeDetector?.props.onStartShouldSetResponder()).toBe(false);
+      // The button marks the tap as RNGH-handled without claiming it...
+      expect(button.props.onStartShouldSetResponderCapture()).toBe(false);
+      // ...so the logical responder claims it and the mark is consumed.
       expect(scrollViewResponder?.props.onStartShouldSetResponder()).toBe(true);
       expect(scrollViewResponder?.props.onStartShouldSetResponder()).toBe(
         false
@@ -281,13 +285,13 @@ describe('[API v3] Components', () => {
 
       await act(flushImmediate);
 
-      const nativeDetector = getNativeDetector(UNSAFE_getAllByType);
       const scrollViewResponder = getScrollViewResponder(UNSAFE_getAllByType);
+      const button = screen.getByTestId('pressable');
 
       // Outside of 'handled' mode the logical responder view is not rendered
       // at all — the responder event can never be claimed on behalf of RNGH.
       expect(scrollViewResponder).toBeUndefined();
-      expect(nativeDetector?.props.onStartShouldSetResponder()).toBe(false);
+      expect(button.props.onStartShouldSetResponderCapture()).toBe(false);
     });
 
     test('handles responder event passed through NativeDetector for keyboardShouldPersistTaps handled', async () => {
